+++ ximpol/evt/event.py
@@ -217,6 +217,11 @@ class xEventFile:
         self.event_data = self.hdu_list['EVENTS'].data
         self.roi_table = self.build_roi_table()
 
+    def close(self):
+        """Close the HDU list.
+        """
+        self.hdu_list.close()
+
     def num_events(self):
         """Return the total number of events in the event file.
         """